+++ noodles-csi/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * csi/writer/index/header/reference_sequence_names: Disallow `NUL` in names.
+
 ## 0.50.0 - 2025-07-12
 
 ### Changed